requirements.txt
flask
python-dotenv
pdfplumber
sentence-transformers==2.7.0
faiss-cpu
numpy
requests
nltk
